Move ClipExtractor to data_extractors
